♻️ extract loadTranslations from webpack.config.js
